# docs/swagger.yaml
basePath: /api
definitions:
  gin.H:
    additionalProperties: {}
    type: object
  models.AuditLog:
    properties:
      action:
        type: string
      correlation_id:
        type: string
      created_at:
        type: string
      details:
        type: string
      id:
        type: string
      ip_address:
        type: string
      resource:
        type: string
      result:
        type: string
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relationships
      user_agent:
        type: string
      user_id:
        type: string
    type: object
  models.BlogPost:
    properties:
      author:
        $ref: '#/definitions/models.User'
      author_id:
        type: string
      category:
        type: string
      content:
        type: string
      created_at:
        type: string
      excerpt:
        type: string
      id:
        type: string
      image:
        type: string
      published_at:
        type: string
      read_time:
        type: integer
      slug:
        type: string
      status:
        description: draft, published, archived
        type: string
      tags:
        description: JSON string of tags
        type: string
      title:
        type: string
      updated_at:
        type: string
      view_count:
        type: integer
    type: object
  models.Session:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      token:
        type: string
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relationships
      user_id:
        type: string
    type: object
  models.SupportTicket:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      priority:
        type: string
      status:
        type: string
      subject:
        type: string
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relationships
      user_id:
        type: string
    type: object
  models.Transaction:
    properties:
      amount:
        type: number
      counterparty_wallet:
        $ref: '#/definitions/models.Wallet'
      counterparty_wallet_id:
        type: string
      created_at:
        type: string
      currency:
        type: string
      description:
        type: string
      id:
        type: string
      idempotency_key:
        type: string
      metadata:
        type: string
      reference_id:
        type: string
      status:
        type: string
      type:
        description: deposit, withdrawal, transfer_in, transfer_out
        type: string
      updated_at:
        type: string
      wallet:
        allOf:
        - $ref: '#/definitions/models.Wallet'
        description: Relationships
      wallet_id:
        type: string
    type: object
  models.User:
    properties:
      audit_logs:
        items:
          $ref: '#/definitions/models.AuditLog'
        type: array
      avatar:
        type: string
      bio:
        type: string
      blog_posts:
        items:
          $ref: '#/definitions/models.BlogPost'
        type: array
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      is_admin:
        type: boolean
      name:
        type: string
      sessions:
        items:
          $ref: '#/definitions/models.Session'
        type: array
      support_tickets:
        items:
          $ref: '#/definitions/models.SupportTicket'
        type: array
      title:
        type: string
      two_factor_enabled:
        type: boolean
      updated_at:
        type: string
      username:
        type: string
      wallets:
        description: Relationships
        items:
          $ref: '#/definitions/models.Wallet'
        type: array
    type: object
  models.Wallet:
    properties:
      balance:
        type: number
      created_at:
        type: string
      currency:
        type: string
      id:
        type: string
      transactions:
        items:
          $ref: '#/definitions/models.Transaction'
        type: array
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relationships
      user_id:
        type: string
      version:
        type: integer
    type: object
...
  routes.ReplyRequest:
    properties:
      message:
        type: string
    required:
    - message
    type: object
  routes.SupportTicketRequest:
    properties:
      message:
        type: string
      priority:
        type: string
      subject:
        type: string
    required:
    - message
    - priority
    - subject
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Secure digital banking platform with JWT authentication, TOTP 2FA,
    multi-currency wallets, atomic transfers, and admin monitoring
  title: SecureWallet API
  version: 1.0.0
paths:
...
  /admin/dashboard:
    get:
      consumes:
      - application/json
      description: Get admin dashboard overview
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Admin dashboard
      tags:
      - admin
  /admin/settings:
    get:
      consumes:
      - application/json
      description: Get current system configuration settings (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Get system settings
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: Update system configuration settings (admin only)
      parameters:
      - description: Settings data
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Save system settings
      tags:
      - admin
  /admin/support/tickets:
    get:
      consumes:
      - application/json
      description: Get all support tickets for admin review
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/gin.H'
            type: array
      security:
      - BearerAuth: []
      summary: Admin list support tickets
      tags:
      - admin
  /admin/support/tickets/{id}/reply:
    post:
      consumes:
      - application/json
      description: Add a reply to a support ticket (admin only)
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: string
      - description: Reply message
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/routes.ReplyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Reply to support ticket
      tags:
      - admin
  /admin/support/tickets/{id}/resolve:
    post:
      consumes:
      - application/json
      description: Mark a support ticket as resolved (admin only)
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Resolve support ticket
      tags:
      - admin
  /admin/transactions:
    get:
      consumes:
      - application/json
      description: Get all system-wide transactions for admin monitoring
      parameters:
      - default: 100
        description: Number of transactions to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/gin.H'
            type: array
      security:
      - BearerAuth: []
      summary: Admin list transactions
      tags:
      - admin
  /admin/users:
    get:
      consumes:
      - application/json
      description: Get all users for admin management
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Admin list users
      tags:
      - admin
  /admin/users/{id}/disable:
    post:
      consumes:
      - application/json
      description: Disable a user account (admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Disable user
      tags:
      - admin
  /admin/users/{id}/enable:
    post:
      consumes:
      - application/json
      description: Enable a previously disabled user account (admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Enable user
      tags:
      - admin
...
  /support/tickets:
    get:
      consumes:
      - application/json
      description: Get all support tickets for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SupportTicket'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: List support tickets
      tags:
      - support
    post:
      consumes:
      - application/json
      description: Create a new support ticket for the authenticated user
      parameters:
      - description: Support ticket data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/routes.SupportTicketRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SupportTicket'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Create support ticket
      tags:
      - support
...
  /transactions:
    get:
      consumes:
      - application/json
      description: Get all transactions for the authenticated user's wallet
      parameters:
      - default: 50
        description: Number of transactions to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Transaction'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: List transactions
      tags:
      - transactions
    post:
      consumes:
      - application/json
      description: Create a new transaction record
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Create transaction
      tags:
      - transactions
...
  /users:
    get:
      consumes:
      - application/json
      description: Get all users with pagination (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/gin.H'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user account (admin only)
      parameters:
      - description: User creation data
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Create user
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    description: Enter the bearer token string with the "Bearer " prefix, e.g. "Bearer
      eyJhbGciOiJIUzI1NiIs..."
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"